Consolidate RunSetService tests into single file
Merge addRunsToRunSet, createRunSetWithRuns, createRunsForRunSet, and
mergeRunSets test files into runSetService.test.ts with shared setup.
